refactors nypr mini chrome
there was hard-coded text in the template. this refactors it out so that
consuming apps can pass in whatever they want.

There was also a * selector which we should generally avoid in our CSS.
This patch moves that additional spacing to padding on the containing
`.nypr-mini-chrome-top` element instead of adding a margin to all
child nodes.